+++ core/Util/plZlib.cpp
@@ -18,11 +18,11 @@
 #include "Stream/hsRAMStream.h"
 #include <zlib.h>
 
-bool plZlib::Uncompress(unsigned char* bufOut, unsigned int& bufLenOut, const unsigned char* bufIn, unsigned int bufLenIn) {
+bool plZlib::Uncompress(unsigned char* bufOut, size_t& bufLenOut, const unsigned char* bufIn, size_t bufLenIn) {
     return (::uncompress(bufOut, (uLongf*)&bufLenOut, bufIn, bufLenIn) == Z_OK);
 }
 
-bool plZlib::Compress(unsigned char*& bufOut, unsigned int& bufLenOut, const unsigned char* bufIn, unsigned int bufLenIn) {
+bool plZlib::Compress(unsigned char*& bufOut, size_t& bufLenOut, const unsigned char* bufIn, size_t bufLenIn) {
     bufLenOut = ::compressBound(bufLenIn);
     bufOut = new unsigned char[bufLenOut];
     return (::compress(bufOut, (uLongf*)&bufLenOut, bufIn, bufLenIn) == Z_OK);
